requirements.txt
flask
flask_sqlalchemy
flask_bcrypt
flask_login
mysqlclient
